Ensure that moc doesn't resolve the qreal meta-type id
When cross-compiling, the qreal type can be different on the target
than on the host (e.g. double on host, and float on target). moc is
a host binary, so it shouldn't try to resolve the type id of qreal,
but instead always output "QMetaType::QReal" (which is just an alias
for QMetaType::Double or QMetaType::Float, depending on the target).

This was a regression introduced in commit
f95181c7bb340744a0ce172e8c5a8fcdc2543297 (new meta-object format);
the special-casing for qreal should have been kept.

Moved the code that generates the type info into its own function so
the logic is shared by generateFunctions() and generateProperties().

Change-Id: I2b76cf063a08ba95a7e6033549452355f67283ac
Reviewed-by: Olivier Goffart <ogoffart@woboq.com>
Reviewed-by: João Abecasis <joao.abecasis@nokia.com>